Convert object rendering fully to floating point math in OpenGL.

+++ libs/s52plib/src/TexFont.h
@@ -54,8 +54,8 @@ public:
   void Delete();
 
   void GetTextExtent(const wxString &string, int *width, int *height);
-  void RenderString(const char *string, int x=0, int y=0, float angle = 0.0);
-  void RenderString(const wxString &string, int x=0, int y=0, float angle = 0.0);
+  void RenderString(const char *string, double x=0, double y=0, float angle = 0.0);
+  void RenderString(const wxString &string, double x=0, double y=0, float angle = 0.0);
   bool IsBuilt() { return m_built; }
   void SetColor(wxColor &color) { m_color = color; }
   void PrepareShader(int width, int height, double rotation);
@@ -77,7 +77,7 @@ private:
   int m_maxglyphh;
   bool m_built;
 
-  float m_dx, m_dy;
+  double m_dx, m_dy;
   float m_angle;
   int m_vpwidth, m_vpheight;
